% Auto-generated: do not edit by hand
\name{pPrimeiroComponente}

\alias{pPrimeiroComponente}

\title{PrimeiroComponente component}

\description{

}

\usage{
pPrimeiroComponente(myDiv=NULL)
}

\arguments{
\item{myDiv}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

