% Auto-generated: do not edit by hand
\name{pTextInput}

\alias{pTextInput}

\title{TextInput component}

\description{

}

\usage{
pTextInput(id=NULL, label=NULL, p=NULL, value=NULL)
}

\arguments{
\item{id}{Character. }

\item{label}{Character. }

\item{p}{Character. }

\item{value}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

